Added a better way to configure the default interpolation function.
